+++ custom_components/alphaess_modbus/number.py
@@ -63,7 +63,6 @@ class AlphaESSNumber(RestoreEntity, NumberEntity):
         self._entry = entry
         self._reg = reg
         self._attr_unique_id = f"{entry.entry_id}_{reg.key}"
-        self._attr_name = reg.name
         self._attr_translation_key = reg.key
         self._attr_native_step = reg.step
         self._attr_native_unit_of_measurement = reg.unit