+++ custom_components/alphaess_modbus/select.py
@@ -46,7 +46,6 @@ class AlphaESSSelect(RestoreEntity, SelectEntity):
         self._entry = entry
         self._reg = reg
         self._attr_unique_id = f"{entry.entry_id}_{reg.key}"
-        self._attr_name = reg.name
         self._attr_translation_key = reg.key
         self._attr_options = reg.options
         self._attr_icon = reg.icon